pyproject.toml
[project]
name = "todoesvan"
version = "0.1.0"
description = "A simple Flet-based todo app"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "FEEH", email = "estividivi2016@hotmail.com" }
]

# Producción
dependencies = [
    "flet==0.28.3",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[tool.flet]
org = "com.mycompany"
product = "todoesvan"
company = "Flet"
copyright = "Copyright (C) 2025 by Flet"

[tool.flet.app]
path = "src"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]  # Errores, fallas, advertencias, import sorting
ignore = ["E501"]              # (opcional) Ignorar long lines, porque black ya formatea
target-version = "py39"

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
strict = true
